// src/entities/prize/model/prize-ids.type.ts
export enum EPrizeIds {
  HOODIE = 'hoodie',
  DOG_TOY = 'dog_toy',
  SOCKS = 'socks',
  BROOM = 'broom',
  PROMO_CODE = 'promo_code',
  KEY_CHAIN = 'key_chain',
  ENTER_BUTTON = 'enter_button',
  KEYBOARD_STICKER = 'keyboard_sticker',
  CLEANING_KIT = 'cleaning_kit',
  PHONE_STAND = 'phone_stand',
  MOUSE_PAD = 'mouse_pad',
  NOTEBOOK = 'notebook',
  CAP = 'cap',
}